requirements.txt
aiobotocore==2.5.2
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
bokeh==3.2.1
botocore==1.29.161
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
contourpy==1.1.0
croniter==1.4.1
cycler==0.11.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.1
docker==6.1.3
docstring-parser==0.15
einops==0.6.1
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonargparse==4.22.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lightning==2.0.5
lightning-api-access==0.0.5
lightning-cloud==0.5.37
lightning-fabric==2.0.5
lightning-utilities==0.9.0
linkify-it-py==2.0.2
lit==16.0.6
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
panel==1.2.0
param==1.13.0
Pillow==9.5.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.3.2
PyYAML==6.0.1
readchar==4.0.5
redis==4.6.0
referencing==0.30.0
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
s3fs==2023.6.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
starsessions==1.3.0
streamlit==1.24.1
sympy==1.12
tenacity==8.2.2
tensorboardX==2.6.1
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typeshed-client==2.3.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.23.1
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
wrapt==1.15.0
xyzservices==2023.7.0
yarl==1.9.2
zipp==3.16.2
